Big code refactor. I nearly everything as it should be format and structure wise. I did this because I was encountering some header file problems so isolating that to a specific source file was very helpful.
Now, the program will change the drive colors which inadvertently resets my own adapater so I've incorporated this as a feature. The idea that the adapter can reset should prove useful in case the adversary is trying to crack the encryption and perhaps has the private key of one or the other ends - for HDCP version 2.x.